- Moved prompt from ollama/ollama.go into commit/prompt.go - Split helpers into logical file names:   - actions.go        # Actions (openInEditor, saveDraft, executeCommit)   - user_prompts.go   # User interaction prompts (choicePrompt, customInputPrompt)   - flow.go           # Commit flow orchestration + CreateCommitMessage router   - cmd.go            # CLI command handler (Main entry point)
