+++ ollama/ollama.go
@@ -198,29 +198,7 @@ func CheckModelExists(baseURL, modelName string) error {
 	return fmt.Errorf(errorMsg)
 }
 
-func CreateCommitMessage(gitDiff string, userConfig *config.UserConfig) (string, error) {
-	prompt := buildCommitPrompt(gitDiff, userConfig)
-	return Main(prompt)
-}
-
-func buildCommitPrompt(gitDiff string, userConfig *config.UserConfig) string {
-	prompt := "You are a commit message generator. Generate a clear, concise commit message based on the following git diff.\n\n"
-
-	prompt += "IMPORTANT: Output ONLY the commit message text. Do not include any explanations, descriptions, or meta-commentary about the commit message.\n\n"
-
-	if userConfig.UseConventional {
-		prompt += "Format: Use Conventional Commits format (type(scope): description)\n"
-	}
-
-	if userConfig.UseEmoji {
-		prompt += "Style: Include appropriate emoji prefixes\n"
-	}
-
-	prompt += fmt.Sprintf("Tone: %s\n", userConfig.Tone)
-	prompt += fmt.Sprintf("Length: %s\n", userConfig.Length)
-
-	prompt += "\nGit diff:\n" + gitDiff + "\n\n"
-	prompt += "Output the commit message now (no other text):"
-
-	return prompt
+func CreateCommitMessage(gitDiff string, userConfig *config.UserConfig, buildPrompt func(string, *config.UserConfig) string) (string, error) {
+	promptText := buildPrompt(gitDiff, userConfig)
+	return Main(promptText)
 }
